Julian

+++ client.c
@@ -6,6 +6,8 @@ void	send_char(int pid, char a)
 	int i;
 
 	i = 7;
+	if(pid == -1 || pid == 0)
+		exit(1);
 	if (kill(pid, 0) == -1)
 		exit(1);
 	while (i >= 0)
@@ -14,42 +16,28 @@ void	send_char(int pid, char a)
 			kill(pid, SIGUSR1);	
 		else
 			kill(pid, SIGUSR2);	
-		usleep(100);
+		usleep(1000);
 		i--;
 	}
 }
 
 void	send_str(char **argv)
 {
+	int	pid;
+
+	pid = ft_atoi(argv[1]);
 	printf("s: send %s\n", argv[2]);
 	while (*argv[2])
 	{
-		send_char(atoi(argv[1]), *argv[2]);
+		send_char(pid, *argv[2]);
 		++(argv[2]);
 	}
-}
-
-void	send_strlen(size_t	len)
-{
-	printf("strlen %ld\n", len);
-	len = 18446744073709551615;
-	int i = 64;
-	while (i >= 0)
-	{
-		if (len >> i & 1)
-			write(1, "1 ", 2);
-		else
-			write(1, "0 ", 2);
-		printf("%d\n", i);
-		usleep(100);
-		i--;
-	}
+	send_char(pid, '\n');
 }
 
 int main(int argc, char **argv)
 {
 	if (argc != 3)
 		exit(1);
-	send_strlen(ft_strlen(argv[2]));
 	send_str(argv);
 }
